moves simulation utils to simcore.utils and adds core.util.get_or_create_system_user()
