+++ SimWorks/core/utils.py
@@ -1,12 +1,11 @@
 import hashlib
 import logging
 import os
-import random
 
 from django.contrib.auth import get_user_model
-from faker import Faker
+from django.core.exceptions import ImproperlyConfigured
 
-fake = Faker()
+from accounts.models import UserRole
 
 # A unique sentinel to detect if a default value was provided.
 _SENTINEL = object()
@@ -30,7 +29,24 @@ def check_env(var_name, default=_SENTINEL):
         error_msg = (
             f"{var_name} not found! Did you set the environment variable {var_name}?"
         )
-        # raise ImproperlyConfigured(error_msg)
+        raise ImproperlyConfigured(error_msg)
+
+def get_or_create_system_user():
+    """
+    Returns the singleton System user for AI or system-generated actions.
+    Creates the user and the 'System' UserRole if they do not exist.
+    """
+    User = get_user_model()
+    role, _ = UserRole.objects.get_or_create(title="System")
+    system_user, _ = User.objects.get_or_create(
+        username="System",
+        defaults={
+            "first_name": "System",
+            "is_active": False,
+            "role": role,
+        }
+    )
+    return system_user
 
 
 class AppColorFormatter(logging.Formatter):
@@ -48,58 +64,6 @@ class AppColorFormatter(logging.Formatter):
         record.name = f"{color}{record.name}{self.RESET}" if color else record.name
         return super().format(record)
 
-
-def generate_fake_name() -> str:
-    return fake.name()
-
-
-def randomize_display_name(full_name: str) -> str:
-    parts = full_name.strip().split()
-    if len(parts) < 2:
-        return full_name  # fallback
-
-    first, last = parts[0], parts[1]
-    initials = f"{first[0]}{last[0]}"
-    options = [
-        full_name,
-        f"{first} {last}",
-        f"{first[0]}. {last}",
-        f"{first} {last[0]}.",
-        f"{first.lower()} {last.lower()}",
-        f"{first.lower()} {last[0].lower()}",
-        f"{initials}",
-        f"{initials.upper()}",
-        f"{first.capitalize()} {last[0].upper()}.",
-        f"{first[0].upper()}.{last[0].upper()}.",
-    ]
-    return random.choice(options)
-
-
-def get_user_initials(user) -> str:
-    """
-    Returns initials for a User object.
-    - If first_name and last_name are set: returns first letters of each
-    - If only username is set: returns first letter
-    - If username is numeric: returns 'Unk'
-    """
-
-    User = get_user_model()
-    if type(user) == str:
-        try:
-            user = User.objects.get(username=user)
-        except User.DoesNotExist:
-            raise Exception(f"Error! get_user_initials: Username {user} not found")
-
-    if (
-        hasattr(user, "first_name")
-        and hasattr(user, "last_name")
-        and user.first_name
-        and user.last_name
-    ):
-        return f"{user.first_name[0]}{user.last_name[0]}".upper()
-    elif hasattr(user, "username") and user.username and not user.username.isnumeric():
-        return user.username[0].upper()
-    return "Unk"
 
 def get_system_user(name="System", **defaults):
     """
